Update evaluation results backup
The corpus used to generate the example evaluation results has been
changed significantly with regards to the division of some main
sections. This refinement of the training data could cause the program
to perform differently - both in a bad or good way. Therefore it has
been run again with the new corpus as training data.
To give the latest evaluation values of the classification program,
update the evaluation files with the latest output.